requirements.txt
streamlit
groq
python-dotenv
matplotlib
beautifulsoup4
